requirements.txt

pydotplus==2.0.2
pytesseract==0.3.13
gTTS==2.5.3
pdf2image== 1.17.0
pillow==10.3.0
python-docx==1.1.2
Levenshtein==0.26.0
opencv==4.10.0
tqdm==4.66.5
opencv-python==4.10.0.84
jupyter==6.5.7
